Allow install of specific (official) versions (#5)
The methods for installing official and unofficial builds have been aligned. It is now possible to install specific official versions with the "--version" options. "--unoffical --version <version>" can be used to lookup the version as well in unofficial builds.

When using "--unofficial" without "--version" options, the latest available version for the host system architecture among official and unofficial builds is installed.

Signed-off-by: MichaIng <micha@dietpi.com>